Web Inspector: ASSERTION FAILED: decodedLength >= dataLength at WebCore::NetworkResourcesData::ResourceData::decodeDataToContent() https://bugs.webkit.org/show_bug.cgi?id=213271 <rdar://problem/64168350>
Reviewed by Brian Burg.

Remove the invalid `ASSERT(decodedLength >= dataLength)` as it's very possible for decoded
content to be smaller than encoded content (e.g. something gzipped).

Use `String::sizeInBytes` instead of `StringImpl::sizeInBytes` as the latter also includes
`sizeof(*this)`, which is not really part of the resource's size, as it's really more of an
implementation detail.

* inspector/NetworkResourcesData.h:
* inspector/NetworkResourcesData.cpp:
(WebCore::NetworkResourcesData::ResourceData::removeContent):
(WebCore::NetworkResourcesData::ResourceData::decodeDataToContent):
(WebCore::NetworkResourcesData::setResourceContent):
(WebCore::NetworkResourcesData::maybeDecodeDataToContent):
(WebCore::contentSizeInBytes): Deleted.


Canonical link: https://commits.webkit.org/226148@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263231 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/inspector/NetworkResourcesData.h
@@ -99,7 +99,7 @@ public:
         bool hasData() const { return m_dataBuffer; }
         size_t dataLength() const;
         void appendData(const char* data, size_t dataLength);
-        size_t decodeDataToContent();
+        unsigned decodeDataToContent();
 
         String m_requestId;
         String m_loaderId;
